Convert decoder limits to float instead of int, to allow more precision
The (about to be submitted) emonTx decoder really needs to be allowed to
specify a limit of 20.3µs, and neither '50 samples' (20µs) nor
'51 samples' (24µs) will suffice.

+++ include/rtl_433.h
@@ -63,9 +63,9 @@ struct protocol_state {
     unsigned int modulation;
 
     /* pwm limits */
-    int short_limit;
-    int long_limit;
-    int reset_limit;
+    float short_limit;
+    float long_limit;
+    float reset_limit;
     char *name;
     unsigned long demod_arg;
 };
